Split upgrade tests into sig-owned directories
