+++ test/e2e/cloud/gcp/network/framework.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2019 The Kubernetes Authors.
+Copyright 2021 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,12 +14,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package gcp
+package network
 
-import (
-	// ensure these packages are scanned by ginkgo for e2e tests
-	_ "k8s.io/kubernetes/test/e2e/cloud/gcp/apps"
-	_ "k8s.io/kubernetes/test/e2e/cloud/gcp/auth"
-	_ "k8s.io/kubernetes/test/e2e/cloud/gcp/network"
-	_ "k8s.io/kubernetes/test/e2e/cloud/gcp/node"
-)
+import "github.com/onsi/ginkgo"
+
+// SIGDescribe annotates the test with the SIG label.
+func SIGDescribe(text string, body func()) bool {
+	return ginkgo.Describe("[sig-network] "+text, body)
+}
